Rename ProjectEnvironmentDefaultStrategy to LegacyProjectEnvironmentDefaultStrategy
